% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_factor_labels.R
\name{apply_factor_labels}
\alias{apply_factor_labels}
\title{apply_factor_labels}
\usage{
apply_factor_labels(data, ids = c("subject_label", "event_sequence"))
}
\arguments{
\item{data}{An data set existing in the user's global environment for
which to apply the factor labels}

\item{ids}{The ID variables used to merge the split factor levels back to the original
dataset, Default: c("subject_label", "event_sequence")}
}
\value{
A data set that converts character variables to factors
}
\description{
Apply factor labels to an existing data set from the ABC-DS codebook
}
\details{
Apply factor labels to an existing data set from the ABC-DS codebook.
This function must be used before the \code{apply_labels} function or the labels
will be removed from the data set.
}
