% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_kbit2_score.R
\name{calculate_kbit2_score}
\alias{calculate_kbit2_score}
\title{calculate_kbit2_score}
\usage{
calculate_kbit2_score(
  age_at_visit,
  kbit2verbkraw,
  kbit2ridraw,
  kbit2vnonvraw,
  print_kbit = TRUE
)
}
\arguments{
\item{age_at_visit}{Age of the participant at visit}

\item{kbit2verbkraw}{Verbal Knowledge Raw Score}

\item{kbit2ridraw}{Riddles Raw Score}

\item{kbit2vnonvraw}{Nonverbal Raw Score}

\item{print_kbit}{Print the KBIT-2 results to the console, Default: TRUE}
}
\value{
A list containing the verbal and nonverbal standard scores, intelligent quotients,
}
\description{
Calculates intelligence quotient and age equivalents from the
KBIT-2 score from look-up tables using the age of the participant and
their verbal and nonverbal scores.
}
\details{
Calculates intelligence quotient and age equivalents from the
KBIT-2 score from look-up tables using the age of the participant and
their verbal and nonverbal scores.
}
\seealso{
\code{\link[dplyr]{filter}}, \code{\link[dplyr]{select}}
}
