% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{deploy_kbit2_app}
\alias{deploy_kbit2_app}
\title{deploy_kbit2_app}
\usage{
deploy_kbit2_app()
}
\value{
Launches the KBIT-2 Calculator Shiny app
}
\description{
Deploys the KBIT-2 Calculator Shiny app on a local host server
}
\details{
Deploys the KBIT-2 Calculator Shiny app on a local host server
}
