% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_kbit2_score.R
\name{print_kbit2_results}
\alias{print_kbit2_results}
\title{print_kbit2_results}
\usage{
print_kbit2_results(res)
}
\arguments{
\item{res}{Results from the \code{calculate_kbit2_score} function}
}
\description{
Prints out the KBIT-2 results from the \code{calculate_kbit2_score} function.
}
\details{
Prints out the KBIT-2 results from the \code{calculate_kbit2_score} function.
}
