% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_medications.R
\name{read_medications_old}
\alias{read_medications_old}
\title{read_medications_old}
\usage{
read_medications_old(
  directory = NULL,
  person = c("participants", "controls"),
  event_sequence = NULL,
  categorize = FALSE
)
}
\arguments{
\item{directory}{A path name containing the downloaded ABC-DS data from the
University of South Carolina Laboratory of Neuro Imaging's (LONI) Image
and Data Archive}

\item{person}{The specific group for which to read in the data, Default: c("participants", "controls")}

\item{event_sequence}{Harmonized event sequence referring to the data collection time point, Default: NULL}

\item{categorize}{Add medication category counts (e.g., number of medications per category), Default: FALSE}
}
\value{
A data frame of the medication data of the participants or controls or
a list containing the medication data of the participants and controls.
}
\description{
Reads the file containing the medication and health history data
from the directory for the participants and controls.
}
\details{
Reads the file containing the medication and health history worksheet
from the directory for the participants and controls. End users have an option
to add medication category counts using the \code{categorize} argument or return the
medication data at a specific time point using \code{event_sequence}. When \code{categorize = TRUE},
the function adds columns counting the number of medications in each category.
Multiple medications may be pipe-delimited (e.g., "6|9") in a single cell.
}
