% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abcds-package.R
\docType{package}
\name{abcds-package}
\alias{abcds-package}
\title{abcds: Eases the Reporting and Analysis of ABC-DS Data}
\description{
Eases the reporting and analysis of data from the Alzheimer's Biomarker Consortium - Down Syndrome (ABC-DS) Study.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bhelsel/abcds}
  \item Report bugs at \url{https://github.com/bhelsel/abcds/issues}
}

}
\author{
\strong{Maintainer}: Brian C. Helsel \email{bhelsel@kumc.edu}

}
\keyword{internal}
