% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abcds-package.R
\name{abcds}
\alias{abcds}
\title{abcds: Eases the Reporting and Analysis of ABC-DS Data}
\description{
Eases the reporting and analysis of data from the
Alzheimer's Biomarker Consortium - Down Syndrome (ABC-DS) Study. The
}
\section{abcds functions}{


\code{\link{add_latency}}

\code{\link{calculate_kbit2_score}}

\code{\link{read_apoe}}

\code{\link{read_cognitive}}

\code{\link{read_demographics}}

\code{\link{read_health_history}}

\code{\link{read_neuro_exam}}
}

