% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_latency.R
\name{add_latency}
\alias{add_latency}
\title{FUNCTION_TITLE}
\usage{
add_latency(
  directory,
  data = NULL,
  person = c("participants", "controls"),
  visit = c("clinical", "amy", "tau", "fdg", "mri", "csf")
)
}
\arguments{
\item{directory}{PARAM_DESCRIPTION}

\item{data}{PARAM_DESCRIPTION, Default: NULL}

\item{person}{PARAM_DESCRIPTION, Default: c("participants", "controls")}

\item{visit}{PARAM_DESCRIPTION, Default: c("clinical", "amy", "tau", "fdg", "mri", "csf")}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
