% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_kbit2_score.R
\name{calculate_kbit2_score}
\alias{calculate_kbit2_score}
\title{calculate_kbit2_score}
\usage{
calculate_kbit2_score(
  age_at_visit,
  kbit2verbkraw,
  kbit2ridraw,
  kbit2vnonvraw,
  subject_label = NULL,
  add_premorbid_id = FALSE,
  print_kbit = TRUE,
  doParallel = TRUE
)
}
\arguments{
\item{age_at_visit}{Age of the participant at visit}

\item{kbit2verbkraw}{Verbal Knowledge Raw Score}

\item{kbit2ridraw}{Riddles Raw Score}

\item{kbit2vnonvraw}{Nonverbal Raw Score}

\item{subject_label}{A subject ID number is required if calculating multiple KBIT-2 scores, Default: NULL}

\item{add_premorbid_id}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} Add a premorbid intellectual disability level using only the KBIT-2 score.
This argument is currently set as FALSE for the default until the premorbid intellectual disability
calculation is better understood, Default: FALSE}

\item{print_kbit}{Print the KBIT-2 results to the console if calculating a single KBIT-2 score, Default: TRUE}

\item{doParallel}{Use parallel processing to speed up the calculation of multiple KBIT-2 scores, Default: TRUE}
}
\value{
A list containing the verbal and nonverbal standard scores, intelligent quotients,
}
\description{
Calculates intelligence quotient and age equivalents from the
KBIT-2 score from look-up tables using the age of the participant and
their verbal and nonverbal scores.
}
\details{
Calculates intelligence quotient and age equivalents from the
KBIT-2 score from look-up tables using the age of the participant and
their verbal and nonverbal scores.
}
\seealso{
\code{\link[dplyr]{filter}}, \code{\link[dplyr]{select}}
}
