% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_apoe.R
\name{read_apoe}
\alias{read_apoe}
\title{FUNCTION_TITLE}
\usage{
read_apoe(
  directory,
  person = c("participants", "controls"),
  include_demographics = FALSE
)
}
\arguments{
\item{directory}{PARAM_DESCRIPTION}

\item{person}{PARAM_DESCRIPTION, Default: c("participants", "controls")}

\item{include_demographics}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
