% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_abcds.R
\name{read_athropometrics}
\alias{read_athropometrics}
\title{Read anthropometric data}
\usage{
read_athropometrics(
  directory = NULL,
  controls = FALSE,
  add_demographics = include_demographics,
  col_types = default_cols
)
}
\arguments{
\item{directory}{Character string. Path to the directory containing data files.
If NULL, uses the path set by \code{\link[=set_abcds_directory]{set_abcds_directory()}}. Default is NULL.}

\item{controls}{Logical. If TRUE, reads control data; if FALSE, reads
participant data. Default is FALSE.}

\item{add_demographics}{Logical. If TRUE, merges demographic data
(age, gender, race, ethnicity) with the returned data. Default is TRUE.}

\item{col_types}{Column specification for \code{\link[readr:read_delim]{readr::read_csv()}}. If NULL,
column types will be guessed.}
}
\value{
A tibble containing height, weight, and body mass index optionally merged with
demographics.
}
\description{
Reads height and weight data from the physical exam data files from the ABCDS study
and converts height and weight to metric before calculating and returning body mass index.
}
\examples{
\dontrun{
# Get height, weight, and body mass index data for participants
anthro_data <- read_athropometrics()
}
}
