% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cognitive.R
\name{read_cognitive}
\alias{read_cognitive}
\title{read_cognitive}
\usage{
read_cognitive(
  directory,
  person = c("participants", "controls"),
  include_demographics = FALSE
)
}
\arguments{
\item{directory}{A path name containing the downloaded ABC-DS data from the
University of South Carolina Laboratory of Neuro Imaging's (LONI) Image
and Data Archive}

\item{person}{The specific group for which to read in the data, Default: c("participants", "controls")}

\item{include_demographics}{An optional parameter to merge demographic information
before returning the data, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Reads the files containing the cognitive function results from the directory
for the participants and controls. The cognitive function assessments are identified as those
starting with \code{cog} using the \code{dd_crf_name} variable in the data dictionary.
}
\details{
Reads the files containing the cognitive function results from the
directory for the participants and controls. The cognitive function
assessments are identified as those starting with \code{cog} using the
\code{dd_crf_name} variable in the data dictionary. End users also have an
option to include the demographics with the \code{include_demographics} argument.
}
