% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_demographics.R
\name{read_demographics}
\alias{read_demographics}
\title{FUNCTION_TITLE}
\usage{
read_demographics(
  directory,
  person = c("participants", "controls"),
  event_sequence = NULL,
  simplify = FALSE
)
}
\arguments{
\item{directory}{PARAM_DESCRIPTION}

\item{person}{PARAM_DESCRIPTION, Default: c("participants", "controls")}

\item{event_sequence}{PARAM_DESCRIPTION, Default: NULL}

\item{simplify}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
