% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_health_history.R
\name{read_health_history}
\alias{read_health_history}
\title{read_health_history}
\usage{
read_health_history(
  directory,
  person = c("participants", "controls"),
  include_demographics = FALSE
)
}
\arguments{
\item{directory}{A path name containing the downloaded ABC-DS data from the
University of South Carolina Laboratory of Neuro Imaging's (LONI) Image
and Data Archive}

\item{person}{The specific group for which to read in the data, Default: c("participants", "controls")}

\item{include_demographics}{An optional parameter to merge demographic information
before returning the data, Default: FALSE}
}
\value{
A data frame of the health history of the participants or controls or
a list containing the health history of the participants and controls.
}
\description{
Reads the file containing the study participant's health history
from the directory for the participants and controls.
}
\details{
Reads the file containing the study participant's health history
from the directory for the participants and controls. End users also have
an option to include the demographics with the \code{include_demographics} argument.
}
