% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_neuro_exam.R
\name{read_neuro_exam}
\alias{read_neuro_exam}
\title{FUNCTION_TITLE}
\usage{
read_neuro_exam(directory, include_demographics = FALSE, ...)
}
\arguments{
\item{directory}{PARAM_DESCRIPTION}

\item{include_demographics}{PARAM_DESCRIPTION, Default: FALSE}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
